linux-image_update: Add provision to mark current running image bootable
The patch adds provision -v or --verify to image_update utility.
Calling image_update with -v needs to be done after user updates an image
and successfully boots from the updated image. The updated image would be
marked non-bootable until user calls image_update -v which marks the current
running image as bootable. This is to indicate to image selector that the
image update was successful and the updated image is bootable.

Signed-off-by: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>