linux-image_update: Added multiboot register clear support
The A/B boot FW update functionality of Kria is dependent on
the interaction of the Image Selector application and the Image Update
utility. The Image Selector application is positioned in physical memory
to be the first image loaded to auto-select the A/B boot image. As
Image Selector uses the ZU+ multi-boot register to make this selection,
and when a soft reset is issued the multi-boot register is not reset it can
result in user booting the same image again and not jumping to the new/updated
image, today we advise users to issue a POR via push-button. This patch
addresses the SRST scenario.

Signed-off-by: Sharath Kumar Dasari <sharath.kumar.dasari@xilinx.com>